[breaking change] - Refactor 'TrimLeftOf' & 'TrimRightOf' extension methods to internal
Move the methods and tests to the internal extension method files
